Adding CSharp test project.

+++ include/campaign/api/api_layout.h
@@ -1,13 +1,13 @@
 #pragma once
 
 #include <campaign/api/api_macros.h>
-#include <stddef.h>
+#include <stdint.h>
 
 #ifdef __cplusplus
 extern "C"
 {
 #endif
-    CAMPAIGN_API void *campaign_layout_create(size_t descriptorCount);
+    CAMPAIGN_API void *campaign_layout_create(int32_t descriptorCount);
 
     CAMPAIGN_API void campaign_layout_delete(void *layoutPtr);
 #ifdef __cplusplus
